Fixed conversion bug
git-svn-id: https://subversion.assembla.com/svn/makefile/trunk@12 795ce9df-6e23-4760-aeee-2b390c6cafd9

+++ C++/launcher/source/launcher.cpp
@@ -22,7 +22,7 @@ int login(const string &username, const string &password) {
 }
 
 void annoyOperator(long numAlerts) {
-    int alerts = addInts(numAlerts, 2);
+    int alerts = addInts(static_cast<int>(numAlerts), 2);
 
     for(int i = 0; i < alerts; i++) {
         cout << "ALARM!!" << endl;
